Move `default-starters.ts`
